test: split admin skill path tests
