+++ crates/dcc-mcp-gateway/src/gateway/admin/tests.rs
@@ -2917,182 +2917,4 @@ webhooks:
         assert_eq!(body["summary"]["live"].as_u64(), Some(1));
         assert_eq!(body["summary"]["stale"].as_u64(), Some(0));
     }
-
-    // ── /api/skill-paths (CRUD) ─────────────────────────────────────────
-
-    #[tokio::test]
-    async fn test_admin_skill_paths_returns_empty_snapshot() {
-        let (status, body) = body_json(admin_router(), "/api/skill-paths").await;
-        assert_eq!(status, StatusCode::OK);
-        assert!(body["paths"].is_array());
-        assert!(body["paths"].as_array().unwrap().is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_admin_skill_paths_shows_snapshot_entries() {
-        use crate::gateway::SkillPathEntry;
-        let state = make_admin_state().with_skill_paths_snapshot(vec![
-            SkillPathEntry {
-                path: "/opt/skills/maya".into(),
-                source: "cli".into(),
-            },
-            SkillPathEntry {
-                path: "/opt/skills/blender".into(),
-                source: "env:DCC_MCP_SKILL_PATHS".into(),
-            },
-        ]);
-        let router = build_admin_router(state);
-        let (status, body) = body_json(router, "/api/skill-paths").await;
-        assert_eq!(status, StatusCode::OK);
-        let paths = body["paths"].as_array().unwrap();
-        assert_eq!(paths.len(), 2);
-        // The display string pairs a friendly source label with a safe folder
-        // tail (never the absolute local path) so same-source rows stay
-        // distinguishable.
-        assert_eq!(paths[0]["path"], "Cli · skills/maya");
-        assert_eq!(paths[0]["display_path"], "Cli · skills/maya");
-        assert_eq!(paths[0]["source_label"], "Cli");
-        assert_eq!(paths[0]["path_tail"], "skills/maya");
-        assert_eq!(paths[0]["path_redacted"], true);
-        assert_ne!(paths[0]["path"], "/opt/skills/maya");
-        assert_eq!(paths[0]["source"], "cli");
-        assert_eq!(paths[0]["status"], "missing");
-        assert_eq!(paths[1]["path"], "Env var · skills/blender");
-        assert_eq!(paths[1]["display_path"], "Env var · skills/blender");
-        assert_eq!(paths[1]["source_label"], "Env var");
-        assert_eq!(paths[1]["path_tail"], "skills/blender");
-        assert_eq!(paths[1]["path_redacted"], true);
-        assert_eq!(paths[1]["source"], "env:DCC_MCP_SKILL_PATHS");
-        assert!(
-            !body.to_string().contains("/opt/skills"),
-            "skill path payload should be safe to attach to public reports"
-        );
-    }
-
-    #[cfg(feature = "admin-persist-sqlite")]
-    #[tokio::test]
-    async fn test_admin_skill_path_crud_via_api() {
-        use crate::gateway::admin::sqlite_lane::AdminSqliteLane;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let db_path = tmp.path().join("test_crud.sqlite");
-        let lane = AdminSqliteLane::spawn(db_path, 30).expect("spawn lane");
-
-        let reload_calls = Arc::new(AtomicUsize::new(0));
-        let reload_calls_for_hook = reload_calls.clone();
-        let state = make_admin_state()
-            .with_admin_sqlite_lane(Some(lane))
-            .with_skill_paths_reload(Some(Arc::new(move || {
-                reload_calls_for_hook.fetch_add(1, Ordering::SeqCst);
-            })));
-        let router = build_admin_router(state);
-
-        // POST a new skill path
-        let resp = router
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/skill-paths")
-                    .header("content-type", "application/json")
-                    .body(axum::body::Body::from(
-                        serde_json::to_vec(&serde_json::json!({"path": "/tmp/new-skills"}))
-                            .unwrap(),
-                    ))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(reload_calls.load(Ordering::SeqCst), 1);
-
-        // GET should now include the new path
-        let (status, body) = body_json(router.clone(), "/api/skill-paths").await;
-        assert_eq!(status, StatusCode::OK);
-        let paths = body["paths"].as_array().unwrap();
-        let custom: Vec<_> = paths
-            .iter()
-            .filter(|p| p["source"] == "admin_custom")
-            .collect();
-        assert_eq!(custom.len(), 1, "expected 1 custom path, got {paths:?}");
-        assert_eq!(custom[0]["path"], "Admin custom · tmp/new-skills");
-        assert_eq!(custom[0]["source_label"], "Admin custom");
-        assert_eq!(custom[0]["path_redacted"], true);
-        assert!(
-            !body.to_string().contains("/tmp/new-skills"),
-            "custom path payload should not expose absolute local paths"
-        );
-
-        // DELETE the custom path
-        let id = custom[0]["id"]
-            .as_i64()
-            .expect("custom path should have id");
-        let resp = router
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("DELETE")
-                    .uri(format!("/api/skill-paths/{id}"))
-                    .body(axum::body::Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(reload_calls.load(Ordering::SeqCst), 2);
-
-        // GET should no longer include the deleted path
-        let (status, body) = body_json(router, "/api/skill-paths").await;
-        assert_eq!(status, StatusCode::OK);
-        let paths = body["paths"].as_array().unwrap();
-        let custom: Vec<_> = paths
-            .iter()
-            .filter(|p| p["source"] == "admin_custom")
-            .collect();
-        assert!(
-            custom.is_empty(),
-            "expected no custom paths after delete, got {custom:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_admin_skill_path_post_empty_returns_400() {
-        let state = make_admin_state();
-        let router = build_admin_router(state);
-        let resp = router
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/skill-paths")
-                    .header("content-type", "application/json")
-                    .body(axum::body::Body::from(
-                        serde_json::to_vec(&serde_json::json!({"path": ""})).unwrap(),
-                    ))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-    }
-
-    #[tokio::test]
-    async fn test_admin_skill_path_post_without_lane_returns_503() {
-        // AdminState without sqlite lane attached
-        let state = make_admin_state();
-        let router = build_admin_router(state);
-        let resp = router
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/skill-paths")
-                    .header("content-type", "application/json")
-                    .body(axum::body::Body::from(
-                        serde_json::to_vec(&serde_json::json!({"path": "/valid/path"})).unwrap(),
-                    ))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(resp.status(), StatusCode::SERVICE_UNAVAILABLE);
-    }
 }
